#43 Modify Login/LoginPopupView
