+++ SsgSag/Assets/Assets.xcassets/background.imageset/Contents.json
@@ -2,18 +2,18 @@
   "images" : [
     {
       "scale" : "1x",
-      "filename" : "background.png",
-      "idiom" : "universal"
+      "idiom" : "universal",
+      "filename" : "background.png"
     },
     {
       "scale" : "2x",
-      "filename" : "background@2x.png",
-      "idiom" : "universal"
+      "idiom" : "universal",
+      "filename" : "background@2x.png"
     },
     {
       "scale" : "3x",
-      "filename" : "background@3x.png",
-      "idiom" : "universal"
+      "idiom" : "universal",
+      "filename" : "background@3x.png"
     }
   ],
   "info" : {
